backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.35
alembic>=1.14.0
psycopg[binary]>=3.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
slowapi>=0.1.9

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
